LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-es-MX-expected.txt

FAIL navigator.language should be es-MX. Was es-XL.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
